refactor: move clerk webhooks
It's been moved directly to the "root" of the `/server` folder.
